(maint) Refactor node creation
Move understanding of how to create nodes into the Node class with a factory
method from_uri. For now make the SSH and WinRM initialize methods take the same
arguments.
