+++ spec/bolt/cli_spec.rb
@@ -69,52 +69,6 @@ describe "Bolt::CLI" do
         cli.parse
       }.to raise_error(Bolt::CLIError, /option --nodes must be specified/)
     end
-
-    describe "with winrm" do
-      it "accepts 'winrm://host:port'" do
-        uri = Bolt::CLI.parse_uri('winrm://neptune:55985')
-        expect(uri.scheme).to eq('winrm')
-        expect(uri.host).to eq('neptune')
-        expect(uri.port).to eq(55985)
-      end
-
-      it "defaults the winrm port to 5985" do
-        uri = Bolt::CLI.parse_uri('winrm://neptune')
-        expect(uri.scheme).to eq('winrm')
-        expect(uri.host).to eq('neptune')
-        expect(uri.port).to eq(5985)
-      end
-    end
-
-    describe "with ssh" do
-      it "accepts 'ssh://host:port'" do
-        uri = Bolt::CLI.parse_uri('ssh://pluto:2224')
-        expect(uri.scheme).to eq('ssh')
-        expect(uri.host).to eq('pluto')
-        expect(uri.port).to eq(2224)
-      end
-
-      it "defaults the ssh port to 22" do
-        uri = Bolt::CLI.parse_uri('ssh://pluto')
-        expect(uri.scheme).to eq('ssh')
-        expect(uri.host).to eq('pluto')
-        expect(uri.port).to eq(22)
-      end
-
-      it "accepts 'host:port' without a scheme" do
-        uri = Bolt::CLI.parse_uri('pluto:2224')
-        expect(uri.scheme).to eq('ssh')
-        expect(uri.host).to eq('pluto')
-        expect(uri.port).to eq(2224)
-      end
-
-      it "defaults the ssh port to 22 without a scheme" do
-        uri = Bolt::CLI.parse_uri('pluto')
-        expect(uri.scheme).to eq('ssh')
-        expect(uri.host).to eq('pluto')
-        expect(uri.port).to eq(22)
-      end
-    end
   end
 
   describe "user" do
